fix empty bundle downloaded
